refactor(cache-schema): consolidate std schema impls
